requirements.txt
streamlit
numpy
pillow
joblib
scikit-learn
